refactor: split components and modal and whatnot
