+++ src/components/ThemeSettings/components/colorStyles.css
@@ -1,0 +1,20 @@
+.vc-usercss-settings-color-swatch {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 69px;
+    height: 50px;
+    box-sizing: border-box;
+    position: relative;
+    border: 1px solid;
+    border-radius: 4px;
+    cursor: pointer;
+}
+
+.vc-usercss-settings-color-swatch-pencil {
+    position: absolute;
+    top: 4px;
+    right: 4px;
+    width: 14px;
+    height: 14px;
+}
